Allow the client to "always ignore" servers with self-signed certificates.
Servers with a certificate chain that is not trusted by the OS (mostly self-signed certificates) can now be "trusted". If the server connects to a server and is notified of a failure in verifying the certificate chain, it will try to match the SHA1 digest of the leaf certificate with a stored digest.  If they match, the user is automatically reconnected to the server in a mode where the certificate chain is not verified.

If the two digests do not match, the user is notified thereof, and can take appropriate action.

For both of these cases, the user can "ignore" the warning, which will cause the client to reconnect in a mode without certificate chain validation. However, the "ignore" option is not remembered, so the same dialog will pop up next time.