MUServerViewController: force PTT to turn off when the app is backgrounded.
This change implements a mechanism in MUServerViewController that will turn off
forceTransmit when the app is sent to the background.

It also ensures that the table view is redrawn (such that the user's talk state is
correctly reflected) before the app is backgrounded. This ensures that when the user
returns to the app, there will not be split second where he can see himself shown as
speaking in the user list, potentially causing a lot of confusion.

Fixes #72.